+++ Sources/KMPObservableBridge/KMPNativeFlow.swift
@@ -4,16 +4,3 @@ public typealias KMPNativeFlow<Output, Failure: Error, Unit> = (
     _ onComplete: @escaping (Failure?, Unit) -> Unit,
     _ onCancelled: @escaping (Failure, Unit) -> Unit
 ) -> () -> Unit
-
-@MainActor
-public protocol KMPNativeObservable: KMPStaticallyObservable {
-    associatedtype KMPObservationOutput
-    associatedtype KMPObservationFailure: Error
-    associatedtype KMPObservationUnit
-
-    var kmpObservationFlow: KMPNativeFlow<
-        KMPObservationOutput,
-        KMPObservationFailure,
-        KMPObservationUnit
-    > { get }
-}