+++ Sources/KMPObservableBridge/KMPValueProperty.swift
@@ -1,5 +1,3 @@
-import SwiftUI
-
 /// A KMP wrapper that exposes its current value through a `value` property.
 @dynamicMemberLookup
 public protocol KMPValueProperty {
@@ -13,12 +11,5 @@ public extension KMPValueProperty {
         dynamicMember keyPath: KeyPath<Value, Member>
     ) -> Member {
         value[keyPath: keyPath]
-    }
-}
-
-public extension Text {
-    init<Property>(_ property: Property)
-    where Property: KMPValueProperty, Property.Value == String {
-        self.init(verbatim: property.value)
     }
 }